fix point data tests
